+++ src/library/SimplePromptbookLibrary.ts
@@ -12,6 +12,7 @@ import type { PromptbookLibrary } from './PromptbookLibrary';
  * Library of promptbooks that groups together promptbooks for an application.
  * This implementation is a very thin wrapper around the Array / Map of promptbooks.
  *
+ * @private createPromptbookLibraryFromJsons
  * @see https://github.com/webgptorg/promptbook#promptbook-library
  */
 export class SimplePromptbookLibrary implements PromptbookLibrary {
@@ -23,7 +24,7 @@ export class SimplePromptbookLibrary implements PromptbookLibrary {
      * @param promptbooks !!!
      *
      * Note: During the construction logic of all promptbooks are validated
-     * Note: It is not recommended to use this constructor directly, use `createPromptbookLibraryFromSources` *(or other variant)* instead
+     * Note: It is not recommended to use this constructor directly, use `createPromptbookLibraryFromJsons` *(or other variant)* instead
      */
     public constructor(...promptbooks: Array<PromptbookJson>) {
         this.library = new Map<string_promptbook_url, PromptbookJson>();