+++ src/library/constructors/createPromptbookLibraryFromJsons.ts
@@ -1,19 +1,17 @@
 import { promptbookStringToJson } from '../../conversion/promptbookStringToJson';
 import type { PromptbookJson } from '../../types/PromptbookJson/PromptbookJson';
-import type { PromptbookString } from '../../types/PromptbookString';
 import { SimplePromptbookLibrary } from '../SimplePromptbookLibrary';
 
 /**
  * Creates PromptbookLibrary from array of PromptbookJson or PromptbookString
  *
- * Note: You can combine `PromptbookString` (`.ptbk.md`) with `PromptbookJson` BUT it is not recommended
  * Note: During the construction syntax and logic of all sources are validated
  *
  * @param promptbookSources
  * @returns PromptbookLibrary
  */
-export async function createPromptbookLibraryFromSources(
-    ...promptbookSources: Array<PromptbookJson | PromptbookString>
+export async function createPromptbookLibraryFromJsons(
+    ...promptbookSources: Array<PromptbookJson>
 ): Promise<SimplePromptbookLibrary> {
     const promptbooks = new Array<PromptbookJson>();
     for (const source of promptbookSources) {
@@ -32,7 +30,6 @@ export async function createPromptbookLibraryFromSources(
     }
     return new SimplePromptbookLibrary(...promptbooks);
 }
-
 
 /**
  * TODO: !!!! [🧠] Library precompilation and do not mix markdown and json promptbooks